Chapter 14/requirements.txt
boto3>=1.34.0
azure-functions>=1.20.0
nbformat>=5.10.0
